backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
motor>=3.7.1
pydantic>=2.12.0
pydantic-settings>=2.7.1
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
